refactor: move minio/database to infrastructure
